Add initial Travis CI
This commit adds in an initial CI. Some CI scripts are pulled from the
IBM/pattern-ci repo. This commit adds CI in the following:
* Linters for Shell and YAML files
* Tests docker-compose yaml and one service endpoint
* Tests REST APIs using Flask unit Tests
  * Modified tests.py to be used in command line
  * Modified tests.py to use local mongo instead
    of recreating mongo containers on each test.
    Mongo database is dropped after each test case.

This commit modified YAML files to pass the YAML linter.

This commit adds __init__.py in web folder so that tests.py can
import it.

This commit modifies The Dockerfile in the flask app. The base image is
now using python alpine image instead of ubuntu. This reduces the total
image size of the app.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>